huff code generation in tree

+++ src/Symbol.kt
@@ -1,4 +1,6 @@
-class Symbol(val symbol: String, val freq: Int) : Comparable<Symbol> {
+import java.math.BigInteger
+
+class Symbol(val symbol: String, val freq: BigInteger) : Comparable<Symbol> {
 
     override fun compareTo(other: Symbol): Int {
         return freq.compareTo(other.freq)
